refactor(gen_reply): make attachment handling a swappable strategy
